Conflicts:
	cypher/src/main/scala/org/neo4j/cypher/internal/commands/ComparablePredicate.scala
	cypher/src/test/scala/org/neo4j/cypher/internal/pipes/SingleShortestPathPipeTest.scala
	embedded-examples/src/test/java/org/neo4j/examples/TraversalTest.java
	kernel/src/main/java/org/neo4j/graphdb/traversal/PruneEvaluator.java
	kernel/src/main/java/org/neo4j/kernel/impl/traversal/TraversalPath.java

+++ kernel/src/main/java/org/neo4j/kernel/impl/batchinsert/PropertyIndexHolder.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2002-2011 "Neo Technology,"
+ * Copyright (c) 2002-2012 "Neo Technology,"
  * Network Engine for Objects in Lund AB [http://neotechnology.com]
  *
  * This file is part of Neo4j.
@@ -26,9 +26,9 @@ import org.neo4j.kernel.impl.util.ArrayMap;
 public class PropertyIndexHolder
 {
     private final ArrayMap<String,Integer> propertyIndexes = 
-        new ArrayMap<String,Integer>( 5, false, false);
+        new ArrayMap<String,Integer>( (byte)5, false, false);
     private final ArrayMap<Integer,String> idToIndex = 
-        new ArrayMap<Integer,String>( 5, false, false);
+        new ArrayMap<Integer,String>( (byte)5, false, false);
     
     PropertyIndexHolder( NameData[] indexes )
     {
