+++ kernel/src/main/java/org/neo4j/kernel/impl/batchinsert/RelationshipTypeHolder.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2002-2011 "Neo Technology,"
+ * Copyright (c) 2002-2012 "Neo Technology,"
  * Network Engine for Objects in Lund AB [http://neotechnology.com]
  *
  * This file is part of Neo4j.
@@ -26,9 +26,9 @@ import org.neo4j.kernel.impl.util.ArrayMap;
 public class RelationshipTypeHolder
 {
     private final ArrayMap<String,Integer> relTypes = 
-        new ArrayMap<String,Integer>( 5, false, false);
+        new ArrayMap<String,Integer>( (byte)5, false, false);
     private final ArrayMap<Integer,String> idToName = 
-        new ArrayMap<Integer,String>( 5, false, false);
+        new ArrayMap<Integer,String>( (byte)5, false, false);
     
     RelationshipTypeHolder( NameData[] types )
     {
