+++ kernel/src/main/java/org/neo4j/kernel/impl/transaction/RWLock.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2002-2011 "Neo Technology,"
+ * Copyright (c) 2002-2012 "Neo Technology,"
  * Network Engine for Objects in Lund AB [http://neotechnology.com]
  *
  * This file is part of Neo4j.
@@ -76,7 +76,7 @@ class RWLock
         new LinkedList<WaitElement>();
 
     private final ArrayMap<Transaction,TxLockElement> txLockElementMap = 
-        new ArrayMap<Transaction,TxLockElement>( 5, false, true );
+        new ArrayMap<Transaction,TxLockElement>( (byte)5, false, true );
 
     private final RagManager ragManager;
