+++ kernel/src/main/java/org/neo4j/kernel/impl/util/ArrayMap.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2002-2011 "Neo Technology,"
+ * Copyright (c) 2002-2012 "Neo Technology,"
  * Network Engine for Objects in Lund AB [http://neotechnology.com]
  *
  * This file is part of Neo4j.
@@ -43,7 +43,7 @@ public class ArrayMap<K,V>
         data = new ArrayEntry[toMapThreshold];
     }
 
-    public ArrayMap( int mapThreshold, boolean threadSafe, boolean shrinkToArray )
+    public ArrayMap( byte mapThreshold, boolean threadSafe, boolean shrinkToArray )
     {
         this.toMapThreshold = (byte)mapThreshold;
         this.useThreadSafeMap = threadSafe;
